feat: added upgrade handler for fee-abs upgrade
