+++ app/upgrades.go
@@ -6,13 +6,13 @@ import (
 	upgradetypes "cosmossdk.io/x/upgrade/types"
 
 	"github.com/pushchain/push-chain-node/app/upgrades"
-	inbound "github.com/pushchain/push-chain-node/app/upgrades/inbound"
+	feeabs "github.com/pushchain/push-chain-node/app/upgrades/fee-abs"
 	"github.com/pushchain/push-chain-node/app/upgrades/noop"
 )
 
 // Upgrades list of chain upgrades
 var Upgrades = []upgrades.Upgrade{
-	inbound.NewUpgrade(),
+	feeabs.NewUpgrade(),
 }
 
 // RegisterUpgradeHandlers registers the chain upgrade handlers
